sql: move code to sqlbase
Moving the go files which were associated with the protos in `sqlbase`, along
with the respective `.go` file and various pieces of code that they were
dependent on. The sqlbase package builds and the sqlbase tests work.
