+++ sql/sqlbase/system.go
@@ -14,7 +14,7 @@
 //
 // Author: Marc Berhault (marc@cockroachlabs.com)
 
-package sql
+package sqlbase
 
 import (
 	"github.com/cockroachdb/cockroach/config"
